pyproject.toml
[tool.poetry]
name = "speech_recognition_bot"
version = "0.1.0"
description = ""
authors = ["balancy <balancy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.1"
python-telegram-bot = "^13.7"
google-cloud-dialogflow = "^2.9.1"
vk-api = "^11.9.6"
loguru = "^0.5.3"

[tool.poetry.dev-dependencies]
black = {version = "^21.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
